refactor: update project navigation structure
