+++ src/App.tsx
@@ -1,13 +1,13 @@
-import { Assets as NavigationAssets } from '@react-navigation/elements';
-import { Asset } from 'expo-asset';
-import * as SplashScreen from 'expo-splash-screen';
-import * as React from 'react';
-import { Navigation } from './navigation';
+import { Assets as NavigationAssets } from "@react-navigation/elements";
+import { Asset } from "expo-asset";
+import * as SplashScreen from "expo-splash-screen";
+import * as React from "react";
+import { Navigation } from "./screens";
 
 Asset.loadAsync([
   ...NavigationAssets,
-  require('./assets/newspaper.png'),
-  require('./assets/bell.png'),
+  require("./assets/newspaper.png"),
+  require("./assets/bell.png"),
 ]);
 
 SplashScreen.preventAutoHideAsync();
@@ -16,10 +16,10 @@ export function App() {
   return (
     <Navigation
       linking={{
-        enabled: 'auto',
+        enabled: "auto",
         prefixes: [
           // Change the scheme to match your app's scheme defined in app.json
-          'helloworld://',
+          "helloworld://",
         ],
       }}
       onReady={() => {
